Create Code for the arduino connected to the step motor
The rotation works quite nicely. However, it is not precise in terms of steps. For reasons unknown to me, the step motor stops a bit too early. As a result, 4096 is not a full rotation. The amount of steps that it misses varied a bit between 2 and 6 (last time I checked). I tried to fix it a bit since then yet it still does miss some steps.