Update Code for the arduino connected to the step motor
Steps and direction of steps (clock/anticlockwise) determinable through the serial monitor.
Step direction, and amount of steps taken into that direction are both displayed in the serial monitor.

Missin: Calculation of nimber of steps into the true degree rotation should be done in the code. Thus, the operator wont have to think of the equivalent step amount for degree that the step motor has to be rotated.